+++ R/utils.R
@@ -18,6 +18,18 @@
   )
 }
 
+.add_speed_limit <- function(url, speed_limit, transport_mode) {
+  if (is.null(speed_limit)) {
+    return(url)
+  } else {
+    paste0(
+      url,
+      ifelse(transport_mode == "pedestrian", "&pedestrian[speed]=", "&vehicle[speedCap]="),
+      speed_limit
+    )
+  }
+}
+
 .add_bbox <- function(url, aoi) {
   bbox <- vapply(sf::st_geometry(aoi), sf::st_bbox, numeric(4))
   .check_bbox(bbox)